Allow disabling MatLog
Users who send logs collected by this program naively believe that
their logs are complete and the developer will be able to see something
in them and fix the problem. However, this program creates a partial
log file. I'm not even talking about dmesg. I don't want to see the logs
from this program. As a rule, they do not contain data about the problem
that the user wants to show.

Let's give the maintainers the opportunity to decide for themselves
whether they need this program or not.

Change-Id: I6e775bf98dd0e0f9e091a664dcf42cee7dc1b870